Fonts edited and improved UX

+++ frontend/src/pages/Admin.jsx
@@ -102,7 +102,8 @@ function Admin() {
 
         // setInfo((prev) => prev.filter(ele => ele.RefugeeID != id))
         // setCampRefs((prev) => prev.filter(ele => ele.RefugeeID != id))
-        deleteMutation(id)
+        const boolDelete = confirm("Do you want to delete the refugee entry");
+        if (boolDelete) deleteMutation(id)
 
         // try {
         //     await AxFetch.delete(`/api/delete/refugee/${id}`)
@@ -130,11 +131,11 @@ function Admin() {
         //   {info && <p>{JSON.stringify(info)}</p>}
         // </div>
         <div className='p-10'>
-            <h1 className='font-bold text-6xl p-10'>The Admin Dashboard</h1>
+            <h1 className='font-bold font-mono text-6xl p-10'>The Admin Dashboard</h1>
             <div className='text-center'>
                 <div className='flex-row'>
-                    <button type="button" class=" font-poppins text-white bg-green-700 hover:bg-green-800 focus:outline-none focus:ring-4 focus:ring-green-300 font-medium rounded-full text-sm px-5 py-2.5 text-center mr-2 mb-2 dark:bg-green-600 dark:hover:bg-green-700 dark:focus:ring-green-800" onClick={(e) => navigate('/create-entry')}>Create Entry</button>
-                    <input placeholder="Enter name" onChange={event => setQuery(event.target.value)} onClick={event => setActive(prev => !prev)} className={/*isActive ? 'border-black border-2 px-7 py-3 w-[80%]' :*/ 'border-black border-y px-7 py-3 w-[60%]'} />
+                    <button type="button" className="text-white bg-green-700 hover:bg-green-800 focus:outline-none focus:ring-4 focus:ring-green-300 font-medium rounded-full text-sm px-5 py-2.5 text-center mr-2 mb-2 dark:bg-green-600 dark:hover:bg-green-700 dark:focus:ring-green-800" onClick={(e) => navigate('/create-entry')}>Create Entry</button>
+                    <input placeholder="Enter name" onChange={event => setQuery(event.target.value)} onClick={event => setActive(prev => !prev)} className={/*isActive ? 'border-black border-2 px-7 py-3 w-[80%]' :*/ 'border-black border-y px-7 py-3 w-[60%] font-mono'} />
                     <FcSearch className='text-center inline-block text-4xl ' />
                 </div>
                 {/* {JSON.stringify(Data)} */}
